[REFACTOR] Extract dummy data into its own project section
Following up on the previous refactor to organize the code
structure, now "fixture" data lives in its own dummy-data
directory. This furthers the code cleanliness and order.
