+++ src/dummy-data/todoItems.js
@@ -1,0 +1,22 @@
+const todoItemsAttributes = [
+  {
+    title: 'Todo Item 1',
+    description: 'This is the first todo item',
+    dueDate: 'Tomorrow',
+    priority: 'High'
+  },
+  {
+    title: 'Todo Item 2',
+    description: 'This is the second todo item',
+    dueDate: 'Next Week',
+    priority: 'Medium'
+  },
+  {
+    title: 'Todo Item 3',
+    description: 'This is the third todo item',
+    dueDate: 'Next Month',
+    priority: 'Low'
+  },
+];
+
+export default todoItemsAttributes;
