+++ src/index.js
@@ -1,3 +1,4 @@
+import todoItemsAttributes from './dummy-data/todoItems';
 import TodoItem from './factories/todoItem';
 
 function component() {
@@ -9,28 +10,6 @@ function component() {
 }
 
 document.body.appendChild(component());
-
-const todoItemsAttributes = [
-  {
-    title: 'Todo Item 1',
-    description: 'This is the first todo item',
-    dueDate: 'Tomorrow',
-    priority: 'High'
-  },
-  {
-    title: 'Todo Item 2',
-    description: 'This is the second todo item',
-    dueDate: 'Next Week',
-    priority: 'Medium'
-  },
-  {
-    title: 'Todo Item 3',
-    description: 'This is the third todo item',
-    dueDate: 'Next Month',
-    priority: 'Low'
-  },
-]
-
 
 const todoItems = todoItemsAttributes.map( (attributes) => TodoItem(attributes));
